Minor details, Spec is not working yet
