+++ templates/rest-boot-api/api/src/main/groovy/groupFolder/restv1/HeroResourceV1.groovy
@@ -1,10 +1,8 @@
 package ${group}.restv1
 
-import gex.javax.ws.rs.Resource
-import org.springframework.beans.factory.annotation.Autowired
 
-import ${group}.domain.v1.Hero
-import ${group}.jaxrs.v1.HeroesResource
+import ${group}.domain.Hero
+import ${group}.jaxrs.v1.HeroResource
 import ${group}.service.HeroService
 
 import gex.javax.ws.rs.Resource
@@ -12,11 +10,13 @@ import gex.jaxrs.ApiResponse
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
 
+import javax.ws.rs.PathParam
+
 import javax.ws.rs.core.Response
 
 @Resource
 @Slf4j
-class HeroesResourceV1 implements HeroesResource{
+class HeroResourceV1 implements HeroResource{
 
   @Autowired
   @Delegate
@@ -26,12 +26,20 @@ class HeroesResourceV1 implements HeroesResource{
   HeroService heroService
 
   @Override
-  Response getHeroes() {
-    ok(heroService.listHeroes())
+  Response getHero(@PathParam("name") String name) {
+    ok(heroService.getHeroByName(name))
   }
 
   @Override
-  Response postHeroes(Hero hero) {
-    created(heroService.createHero(hero))
+  Response putHero(@PathParam("name") String name, Hero hero) {
+    ok(heroService.updateHero(name, hero))
+  }
+
+  @Override
+  Response deleteHero(@PathParam("name") String name) {
+    heroService.deleteHero(name)
+    noContent()
   }
 }
+
+
