+++ templates/rest-boot-api/gradle.properties
@@ -6,7 +6,7 @@ pomDescription=${projectName}
 
 javaVersion=1.8
 
-commonsExceptionsVersion=0.1.2
+commonsExceptionsVersion=0.1.4
 gexJaxrsAstVersion=0.1.15
 gexJaxrsProvidersVersion=0.0.6
 gormHibernateVersion=1.1.0.RELEASE
